change user role and protect the delete functionality

+++ convex/files.ts
@@ -1,6 +1,6 @@
 import { ConvexError, v } from "convex/values";
 import { mutation, MutationCtx, query, QueryCtx } from "./_generated/server";
-import { getUser } from "./users";
+import { access } from "fs";
 
 export const generateUploadUrl = mutation(async (ctx) => {
   const identity = await ctx.auth.getUserIdentity();
@@ -10,15 +10,26 @@ export const generateUploadUrl = mutation(async (ctx) => {
   return await ctx.storage.generateUploadUrl();
 });
 
-async function hasAccessToOrg(
-  ctx: MutationCtx | QueryCtx,
-  tokenIdentifier: string,
-  orgId: string
-) {
-  const user = await getUser(ctx, tokenIdentifier);
+async function hasAccessToOrg(ctx: MutationCtx | QueryCtx, orgId: string) {
+  const identity = await ctx.auth.getUserIdentity();
+  if (!identity) {
+    return null;
+  }
+  const user = await ctx.db
+    .query("users")
+    .withIndex("by_tokenIdentifier", (q) =>
+      q.eq("tokenIdentifier", identity.tokenIdentifier)
+    )
+    .first();
+  if (!user) {
+    return null;
+  }
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
-  return hasAccess;
+    user.orgIds.some(item => item.orgId === orgId) || user.tokenIdentifier.includes(orgId);
+  if (!hasAccess) {
+    return null;
+  }
+  return { user };
 }
 
 export const createFile = mutation({
@@ -29,16 +40,7 @@ export const createFile = mutation({
     type: v.union(v.literal("image"), v.literal("pdf"), v.literal("csv")),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-    if (!identity) {
-      throw new ConvexError("Not authenticated");
-    }
-
-    const hasAccess = await hasAccessToOrg(
-      ctx,
-      identity.tokenIdentifier,
-      args.orgId
-    );
+    const hasAccess = await hasAccessToOrg(ctx, args.orgId);
 
     if (!hasAccess) {
       throw new ConvexError("you do not have access to this org");
@@ -57,17 +59,10 @@ export const getFiles = query({
   args: {
     orgId: v.string(),
     query: v.optional(v.string()),
+    favorite: v.optional(v.boolean()),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-    if (!identity) {
-      return [];
-    }
-    const hasAccess = await hasAccessToOrg(
-      ctx,
-      identity.tokenIdentifier,
-      args.orgId
-    );
+    const hasAccess = await hasAccessToOrg(ctx, args.orgId);
     if (!hasAccess) {
       return [];
     }
@@ -75,15 +70,25 @@ export const getFiles = query({
       .query("files")
       .withIndex("by_orgId", (q) => q.eq("orgId", args.orgId))
       .collect();
-      
+
     const query = args.query;
     if (query) {
-      return files.filter((file) =>
+      files = files.filter((file) =>
         file.name.toLowerCase().includes(query.toLowerCase())
       );
-    } else {
-      return files;
     }
+    if (args.favorite) {
+      const favorite = await ctx.db
+        .query("favorite")
+        .withIndex("by_userId_orgId_fileId", (q) =>
+          q.eq("userId", hasAccess.user._id).eq("orgId", args.orgId)
+        )
+        .collect();
+      files = files.filter((file) =>
+        favorite.some((f) => f.fileId === file._id)
+      );
+    }
+    return files;
   },
 });
 
@@ -92,23 +97,73 @@ export const deleteFile = mutation({
     fileId: v.id("files"),
   },
   async handler(ctx, args) {
-    const identity = await ctx.auth.getUserIdentity();
-
-    if (!identity) {
-      throw new ConvexError("Not authenticated");
-    }
     const file = await ctx.db.get(args.fileId);
     if (!file) {
       throw new ConvexError("file not found");
     }
-    const hasAccess = await hasAccessToOrg(
-      ctx,
-      identity.tokenIdentifier,
-      file.orgId || ""
-    );
+    const hasAccess = await hasAccessToOrg(ctx, file.orgId || "");
     if (!hasAccess) {
       throw new ConvexError("have no access to the organiztion");
     }
+    const isAdmin = hasAccess.user.orgIds.find(org => org.orgId === file.orgId)?.role === "admin";
+    if(!isAdmin) {
+      throw new ConvexError("you do not have access to delete this file");
+    }
     await ctx.db.delete(args.fileId);
+  },
+});
+
+export const toggleFavorite = mutation({
+  args: {
+    fileId: v.id("files"),
+  },
+  async handler(ctx, args) {
+    const file = await ctx.db.get(args.fileId);
+    if (!file) {
+      throw new ConvexError("file not found");
+    }
+    const hasAccess = await hasAccessToOrg(ctx, file.orgId || "");
+    if (!hasAccess) {
+      throw new ConvexError("have no access to the organiztion");
+    }
+    const favorite = await ctx.db
+      .query("favorite")
+      .withIndex("by_userId_orgId_fileId", (q) =>
+        q
+          .eq("userId", hasAccess.user._id)
+          .eq("orgId", file.orgId ?? "")
+          .eq("fileId", file._id)
+      )
+      .first();
+    if (!favorite) {
+      await ctx.db.insert("favorite", {
+        fileId: file._id,
+        orgId: file.orgId ?? "",
+        userId: hasAccess.user._id,
+      });
+    } else {
+      await ctx.db.delete(favorite._id);
+    }
+  },
+});
+
+export const getAllFavorite = query({
+  args: {
+    orgId: v.string(),
+  },
+  async handler(ctx, args) {
+    const hasAccess = await hasAccessToOrg(ctx, args.orgId || "");
+    if (!hasAccess) {
+      throw new ConvexError("have no access to the organiztion");
+    }
+    const favorite = await ctx.db
+      .query("favorite")
+      .withIndex("by_userId_orgId_fileId", (q) =>
+        q
+          .eq("userId", hasAccess.user._id)
+          .eq("orgId", args.orgId ?? "")
+      )
+      .collect()
+    return favorite;
   },
 });
